feat: el motor sherpa se descarga a demanda y sale del paquete
Mismo trato que el motor sonata en las PR #133 y #134, ahora para el
servidor de las voces Kokoro: 40 MB menos en el .msi para todo el mundo,
y solo los baja quien de verdad va a usar estas voces.

Como ya son dos motores con el mismo paquete (.tar.bz2 con firma SHA-256
en la release fija «motores»), el instalador pasa a ser común en vez de
copiarse por tercera vez:

- servicios/motor_manager.py: el trabajo (progreso 0-100, verificación de
  la firma, extracción en streaming, cancelación en todo momento). Sonata
  y sherpa solo ponen sus constantes.
- ui/motor_downloader.py y controller/motor_downloader_controller.py: el
  mismo diálogo y el mismo controlador para los dos, con el título, los
  textos y la recarga de la voz en cada subclase.

El motor se resuelve como el de sonata: 64/sherpa manda si está (árbol de
desarrollo, instalaciones anteriores), engines/sherpa es para las que ya
no lo traen. Se resuelve en cada __init__ del puente y no al importar, así
un motor recién descargado se encuentra sin reiniciar VeTube.

Sin motor, configurar_tts("kokoro") devuelve el respaldo SAPI del instante
en lugar de levantar un servidor que no está: nadie se queda mudo sin
explicación, y config['sistemaTTS'] no se toca. La descarga se ofrece en
los tres sitios de sonata (arranque, Aceptar de los Ajustes, prueba de
voz), y siempre ANTES que el paquete de voces: es pequeño, y sin él las
voces de 334 MB no sonarían igualmente.

pyproject.toml deja de copiar 64/sherpa. La carpeta sigue en el árbol.

i18n: 7 msgid nuevos (los gemelos de los del descargador de sonata), los 6
catálogos al día y recompilados; ru queda sin ellos a propósito, con el
respaldo español, como en la #133.
