+++ TTS/lector.py
@@ -58,6 +58,16 @@ def configurar_tts(lector):
             return crear_respaldo_puente()
         return sonata_handler.piperSpeak()
     elif lector == "kokoro":
+        if not sherpa_handler.sherpa_instalado():
+            # Mismo caso que Piper sin sonata: el motor sherpa tampoco viaja ya
+            # en el build (se descarga desde la release «motores»), así que sin
+            # él levantar el puente dejaría al usuario mudo sin explicación.
+            # Respaldo SAPI del instante, sin tocar config['sistemaTTS'].
+            logger.warning(
+                "El motor sherpa no está instalado: las voces Kokoro hablarán "
+                "por el respaldo SAPI hasta que se descargue."
+            )
+            return crear_respaldo_puente()
         return sherpa_handler.sherpaSpeak()
     elif lector == "edge":
         return edge_handler.edgeSpeak()
