+++ locales/cs/LC_MESSAGES/VeTube.po
@@ -2090,13 +2090,25 @@ msgstr "Pro kanál %s není k dispozici žádná kompatibilní verze. Kanál Sta
 msgid "Instalar el motor de las voces Piper"
 msgstr "Nainstalovat modul hlasů Piper"
 
+#: controller/sherpa_downloader_controller.py:24
+msgid "Instalar el motor de las voces Kokoro"
+msgstr "Nainstalovat modul hlasů Kokoro"
+
 #: ui/sonata_downloader.py:21
 msgid "Las voces Piper necesitan el motor sonata para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
 msgstr "Hlasy Piper potřebují ke svému fungování modul sonata. Stahuje se pouze jednou a zabírá přibližně %d MB."
 
+#: controller/sherpa_downloader_controller.py:29
+msgid "Las voces Kokoro necesitan el motor sherpa para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
+msgstr "Hlasy Kokoro potřebují ke svému fungování modul sherpa. Stahuje se pouze jednou a zabírá přibližně %d MB."
+
 #: controller/sonata_downloader_controller.py:30
 msgid "El motor de las voces Piper ya está instalado en este equipo."
 msgstr "Modul hlasů Piper je již v tomto počítači nainstalován."
+
+#: controller/sherpa_downloader_controller.py:36
+msgid "El motor de las voces Kokoro ya está instalado en este equipo."
+msgstr "Modul hlasů Kokoro je již v tomto počítači nainstalován."
 
 #: controller/sonata_downloader_controller.py:50
 msgid "Descargando el motor de voz..."
@@ -2106,9 +2118,17 @@ msgstr "Stahování hlasového modulu..."
 msgid "El motor de las voces Piper se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
 msgstr "Modul hlasů Piper byl úspěšně nainstalován. Zprávy nyní mohou znít zvoleným hlasem."
 
+#: controller/sherpa_downloader_controller.py:40
+msgid "El motor de las voces Kokoro se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
+msgstr "Modul hlasů Kokoro byl úspěšně nainstalován. Zprávy nyní mohou znít zvoleným hlasem."
+
 #: controller/sonata_downloader_controller.py:95
 msgid "No se pudo instalar el motor de las voces Piper: %s"
 msgstr "Modul hlasů Piper se nepodařilo nainstalovat: %s"
+
+#: controller/sherpa_downloader_controller.py:44
+msgid "No se pudo instalar el motor de las voces Kokoro: %s"
+msgstr "Modul hlasů Kokoro se nepodařilo nainstalovat: %s"
 
 #: servicios/sonata_manager.py:188
 msgid "el paquete descargado no supera la comprobación de integridad. Inténtalo de nuevo."
@@ -2118,6 +2138,14 @@ msgstr "stažený balíček neprošel kontrolou integrity. Zkuste to znovu."
 msgid "Las voces Piper necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
 msgstr "Hlasy Piper potřebují svůj hlasový modul, který v tomto počítači zatím není. Chcete jej stáhnout nyní? Mezitím vás bude provázet systémový hlas."
 
+#: utils/app_utilitys.py:111
+msgid "Las voces Kokoro necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
+msgstr "Hlasy Kokoro potřebují svůj hlasový modul, který v tomto počítači zatím není. Chcete jej stáhnout nyní? Mezitím vás bude provázet systémový hlas."
+
 #: utils/app_utilitys.py:81
 msgid "Falta el motor de las voces Piper"
 msgstr "Chybí modul hlasů Piper"
+
+#: utils/app_utilitys.py:113
+msgid "Falta el motor de las voces Kokoro"
+msgstr "Chybí modul hlasů Kokoro"
