+++ locales/fr/LC_MESSAGES/VeTube.po
@@ -2114,13 +2114,25 @@ msgstr "Aucune version compatible n'est disponible pour le canal %s. Le canal St
 msgid "Instalar el motor de las voces Piper"
 msgstr "Installer le moteur des voix Piper"
 
+#: controller/sherpa_downloader_controller.py:24
+msgid "Instalar el motor de las voces Kokoro"
+msgstr "Installer le moteur des voix Kokoro"
+
 #: ui/sonata_downloader.py:21
 msgid "Las voces Piper necesitan el motor sonata para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
 msgstr "Les voix Piper ont besoin du moteur sonata pour pouvoir parler. Il se télécharge une seule fois et occupe environ %d Mo."
 
+#: controller/sherpa_downloader_controller.py:29
+msgid "Las voces Kokoro necesitan el motor sherpa para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
+msgstr "Les voix Kokoro ont besoin du moteur sherpa pour pouvoir parler. Il se télécharge une seule fois et occupe environ %d Mo."
+
 #: controller/sonata_downloader_controller.py:30
 msgid "El motor de las voces Piper ya está instalado en este equipo."
 msgstr "Le moteur des voix Piper est déjà installé sur cet ordinateur."
+
+#: controller/sherpa_downloader_controller.py:36
+msgid "El motor de las voces Kokoro ya está instalado en este equipo."
+msgstr "Le moteur des voix Kokoro est déjà installé sur cet ordinateur."
 
 #: controller/sonata_downloader_controller.py:50
 msgid "Descargando el motor de voz..."
@@ -2130,9 +2142,17 @@ msgstr "Téléchargement du moteur de voix..."
 msgid "El motor de las voces Piper se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
 msgstr "Le moteur des voix Piper a été installé correctement. Les messages peuvent désormais être lus avec la voix choisie."
 
+#: controller/sherpa_downloader_controller.py:40
+msgid "El motor de las voces Kokoro se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
+msgstr "Le moteur des voix Kokoro a été installé correctement. Les messages peuvent désormais être lus avec la voix choisie."
+
 #: controller/sonata_downloader_controller.py:95
 msgid "No se pudo instalar el motor de las voces Piper: %s"
 msgstr "Impossible d'installer le moteur des voix Piper : %s"
+
+#: controller/sherpa_downloader_controller.py:44
+msgid "No se pudo instalar el motor de las voces Kokoro: %s"
+msgstr "Impossible d'installer le moteur des voix Kokoro : %s"
 
 #: servicios/sonata_manager.py:188
 msgid "el paquete descargado no supera la comprobación de integridad. Inténtalo de nuevo."
@@ -2142,6 +2162,14 @@ msgstr "le paquet téléchargé n'a pas passé la vérification d'intégrité. R
 msgid "Las voces Piper necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
 msgstr "Les voix Piper ont besoin de leur moteur de voix, qui n'est pas encore sur cet ordinateur. Voulez-vous le télécharger maintenant ? En attendant, une voix du système vous accompagnera."
 
+#: utils/app_utilitys.py:111
+msgid "Las voces Kokoro necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
+msgstr "Les voix Kokoro ont besoin de leur moteur de voix, qui n'est pas encore sur cet ordinateur. Voulez-vous le télécharger maintenant ? En attendant, une voix du système vous accompagnera."
+
 #: utils/app_utilitys.py:81
 msgid "Falta el motor de las voces Piper"
 msgstr "Le moteur des voix Piper est manquant"
+
+#: utils/app_utilitys.py:113
+msgid "Falta el motor de las voces Kokoro"
+msgstr "Le moteur des voix Kokoro est manquant"
