+++ locales/id/LC_MESSAGES/VeTube.po
@@ -2113,13 +2113,25 @@ msgstr "Tidak ada versi yang kompatibel untuk saluran %s. Saluran Stable mungkin
 msgid "Instalar el motor de las voces Piper"
 msgstr "Pasang mesin suara Piper"
 
+#: controller/sherpa_downloader_controller.py:24
+msgid "Instalar el motor de las voces Kokoro"
+msgstr "Pasang mesin suara Kokoro"
+
 #: ui/sonata_downloader.py:21
 msgid "Las voces Piper necesitan el motor sonata para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
 msgstr "Suara Piper membutuhkan mesin sonata agar dapat berbicara. Mesin ini hanya diunduh sekali dan berukuran sekitar %d MB."
 
+#: controller/sherpa_downloader_controller.py:29
+msgid "Las voces Kokoro necesitan el motor sherpa para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
+msgstr "Suara Kokoro membutuhkan mesin sherpa agar dapat berbicara. Mesin ini hanya diunduh sekali dan berukuran sekitar %d MB."
+
 #: controller/sonata_downloader_controller.py:30
 msgid "El motor de las voces Piper ya está instalado en este equipo."
 msgstr "Mesin suara Piper sudah terpasang di komputer ini."
+
+#: controller/sherpa_downloader_controller.py:36
+msgid "El motor de las voces Kokoro ya está instalado en este equipo."
+msgstr "Mesin suara Kokoro sudah terpasang di komputer ini."
 
 #: controller/sonata_downloader_controller.py:50
 msgid "Descargando el motor de voz..."
@@ -2129,9 +2141,17 @@ msgstr "Mengunduh mesin suara..."
 msgid "El motor de las voces Piper se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
 msgstr "Mesin suara Piper berhasil dipasang. Pesan sekarang dapat dibacakan dengan suara yang dipilih."
 
+#: controller/sherpa_downloader_controller.py:40
+msgid "El motor de las voces Kokoro se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
+msgstr "Mesin suara Kokoro berhasil dipasang. Pesan sekarang dapat dibacakan dengan suara yang dipilih."
+
 #: controller/sonata_downloader_controller.py:95
 msgid "No se pudo instalar el motor de las voces Piper: %s"
 msgstr "Mesin suara Piper tidak dapat dipasang: %s"
+
+#: controller/sherpa_downloader_controller.py:44
+msgid "No se pudo instalar el motor de las voces Kokoro: %s"
+msgstr "Mesin suara Kokoro tidak dapat dipasang: %s"
 
 #: servicios/sonata_manager.py:188
 msgid "el paquete descargado no supera la comprobación de integridad. Inténtalo de nuevo."
@@ -2141,6 +2161,14 @@ msgstr "paket yang diunduh tidak lolos pemeriksaan integritas. Coba lagi."
 msgid "Las voces Piper necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
 msgstr "Suara Piper membutuhkan mesin suaranya, yang belum ada di komputer ini. Ingin mengunduhnya sekarang? Sementara itu, suara sistem akan menemani Anda."
 
+#: utils/app_utilitys.py:111
+msgid "Las voces Kokoro necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
+msgstr "Suara Kokoro membutuhkan mesin suaranya, yang belum ada di komputer ini. Ingin mengunduhnya sekarang? Sementara itu, suara sistem akan menemani Anda."
+
 #: utils/app_utilitys.py:81
 msgid "Falta el motor de las voces Piper"
 msgstr "Mesin suara Piper belum ada"
+
+#: utils/app_utilitys.py:113
+msgid "Falta el motor de las voces Kokoro"
+msgstr "Mesin suara Kokoro belum ada"
