+++ locales/pl/LC_MESSAGES/VeTube.po
@@ -2111,13 +2111,25 @@ msgstr "Brak zgodnej wersji dla kanału %s. Kanał Stable może nie mieć zgodne
 msgid "Instalar el motor de las voces Piper"
 msgstr "Zainstaluj silnik głosów Piper"
 
+#: controller/sherpa_downloader_controller.py:24
+msgid "Instalar el motor de las voces Kokoro"
+msgstr "Zainstaluj silnik głosów Kokoro"
+
 #: ui/sonata_downloader.py:21
 msgid "Las voces Piper necesitan el motor sonata para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
 msgstr "Głosy Piper potrzebują silnika sonata, aby mogły mówić. Jest pobierany tylko raz i zajmuje około %d MB."
 
+#: controller/sherpa_downloader_controller.py:29
+msgid "Las voces Kokoro necesitan el motor sherpa para poder sonar. Se descarga una sola vez y ocupa %d MB aproximadamente."
+msgstr "Głosy Kokoro potrzebują silnika sherpa, aby mogły mówić. Jest pobierany tylko raz i zajmuje około %d MB."
+
 #: controller/sonata_downloader_controller.py:30
 msgid "El motor de las voces Piper ya está instalado en este equipo."
 msgstr "Silnik głosów Piper jest już zainstalowany na tym komputerze."
+
+#: controller/sherpa_downloader_controller.py:36
+msgid "El motor de las voces Kokoro ya está instalado en este equipo."
+msgstr "Silnik głosów Kokoro jest już zainstalowany na tym komputerze."
 
 #: controller/sonata_downloader_controller.py:50
 msgid "Descargando el motor de voz..."
@@ -2127,9 +2139,17 @@ msgstr "Pobieranie silnika głosu..."
 msgid "El motor de las voces Piper se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
 msgstr "Silnik głosów Piper został pomyślnie zainstalowany. Wiadomości mogą już być odczytywane wybranym głosem."
 
+#: controller/sherpa_downloader_controller.py:40
+msgid "El motor de las voces Kokoro se ha instalado correctamente. Los mensajes ya pueden sonar con la voz elegida."
+msgstr "Silnik głosów Kokoro został pomyślnie zainstalowany. Wiadomości mogą już być odczytywane wybranym głosem."
+
 #: controller/sonata_downloader_controller.py:95
 msgid "No se pudo instalar el motor de las voces Piper: %s"
 msgstr "Nie udało się zainstalować silnika głosów Piper: %s"
+
+#: controller/sherpa_downloader_controller.py:44
+msgid "No se pudo instalar el motor de las voces Kokoro: %s"
+msgstr "Nie udało się zainstalować silnika głosów Kokoro: %s"
 
 #: servicios/sonata_manager.py:188
 msgid "el paquete descargado no supera la comprobación de integridad. Inténtalo de nuevo."
@@ -2139,6 +2159,14 @@ msgstr "pobrany pakiet nie przeszedł kontroli integralności. Spróbuj ponownie
 msgid "Las voces Piper necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
 msgstr "Głosy Piper potrzebują swojego silnika głosu, którego nie ma jeszcze na tym komputerze. Czy chcesz pobrać go teraz? W międzyczasie będzie Ci towarzyszyć głos systemowy."
 
+#: utils/app_utilitys.py:111
+msgid "Las voces Kokoro necesitan su motor de voz, que aún no está en este equipo. ¿Deseas descargarlo ahora? Mientras tanto te acompañará una voz del sistema."
+msgstr "Głosy Kokoro potrzebują swojego silnika głosu, którego nie ma jeszcze na tym komputerze. Czy chcesz pobrać go teraz? W międzyczasie będzie Ci towarzyszyć głos systemowy."
+
 #: utils/app_utilitys.py:81
 msgid "Falta el motor de las voces Piper"
 msgstr "Brakuje silnika głosów Piper"
+
+#: utils/app_utilitys.py:113
+msgid "Falta el motor de las voces Kokoro"
+msgstr "Brakuje silnika głosów Kokoro"
