+++ pyproject.toml
@@ -168,16 +168,16 @@ include_files = [
     ["locales/", "locales/"],
     ["sounds/", "sounds/"],
 
-    # OJO: el contenido de 64/ va enumerado UNO POR UNO, a propósito. El motor
-    # sonata (64/sonata, unos 48 MB) ya no viaja en el build: se descarga a
-    # demanda desde la release «motores». Como la carpeta ya no se copia
-    # entera, TODO lo que se añada dentro de 64/ hay que apuntarlo también
-    # aquí, o no llegará al build sin dar el menor aviso.
+    # OJO: el contenido de 64/ va enumerado UNO POR UNO, a propósito. Los dos
+    # motores de voz (64/sonata, unos 48 MB, y 64/sherpa, unos 40 MB) ya no
+    # viajan en el build: se descargan a demanda desde la release «motores».
+    # Como la carpeta ya no se copia entera, TODO lo que se añada dentro de
+    # 64/ hay que apuntarlo también aquí, o no llegará al build sin dar el
+    # menor aviso.
     ["64/bypass64.exe", "lib/64/bypass64.exe"],
     ["64/libvlc.dll", "lib/64/libvlc.dll"],
     ["64/libvlccore.dll", "lib/64/libvlccore.dll"],
     ["64/plugins/", "lib/64/plugins/"], # VLC
-    ["64/sherpa/", "lib/64/sherpa/"], # motor Kokoro
 
     [".venv/Lib/site-packages/sound_lib/lib/", "sound_lib/lib/"],
 ]
